Добавляет HOC для карты
