+++ src/components/map/map.test.js
@@ -5,7 +5,7 @@ import Map from "./map.jsx";
 
 import testMocks from "../../test-mocks/map.js";
 
-describe.skip(`snapshot test: Map component`, () => {
+describe(`snapshot test: Map component`, () => {
   it(`should render correctly`, () => expect(
       shallow(<Map {...testMocks}/>)
   ).toMatchSnapshot());
