+++ src/test-mocks/map.js
@@ -1,5 +1,3 @@
 export default {
-  blockClassName: `block`,
-  markerCoordinates: [[4, 4]],
-  activeMarkerCoordinates: [[4, 4]]
+  blockClassName: `block`
 };
